chore: Integrate `Channel` from Awala core library
