+++ src/server/routes/connectionParams.routes.spec.ts
@@ -1,14 +1,14 @@
 import type { FastifyInstance } from 'fastify';
 
 import { makeTestPohttpServer } from '../../testUtils/pohttpServer.js';
-import type { InternetEndpointManager } from '../../utilities/awala/InternetEndpointManager.js';
+import type { InternetEndpoint } from '../../utilities/awala/InternetEndpoint.js';
 
 describe('Connection params route', () => {
   const getTestServerFixture = makeTestPohttpServer();
   let server: FastifyInstance;
-  let endpointManager: InternetEndpointManager;
+  let endpoint: InternetEndpoint;
   beforeEach(() => {
-    ({ server, endpointManager } = getTestServerFixture());
+    ({ server, endpoint } = getTestServerFixture());
   });
 
   test('Should respond with connection parameters', async () => {
@@ -16,8 +16,7 @@ describe('Connection params route', () => {
 
     expect(response).toHaveProperty('statusCode', 200);
     expect(response).toHaveProperty('headers.content-type', 'application/vnd.etsi.tsl.der');
-    const internetEndpoint = await endpointManager.getActiveEndpoint();
-    const parameters = await internetEndpoint.getConnectionParams();
+    const parameters = await endpoint.getConnectionParams();
     expect(response.rawPayload).toMatchObject(parameters);
   });
 });
