+++ src/testUtils/server.ts
@@ -2,7 +2,7 @@ import type { FastifyInstance } from 'fastify';
 import type { Connection } from 'mongoose';
 
 import type { ServerMaker } from '../utilities/fastify/ServerMaker.js';
-import type { InternetEndpointManager } from '../utilities/awala/InternetEndpointManager.js';
+import type { InternetEndpoint } from '../utilities/awala/InternetEndpoint.js';
 
 import { makeMockLogging, type MockLogSet } from './logging.js';
 import { configureMockEnvVars, type EnvVarSet, type EnvVarMocker } from './envVars.js';
@@ -14,7 +14,7 @@ export interface TestServerFixture {
   readonly dbConnection: Connection;
   readonly logs: MockLogSet;
   readonly envVarMocker: EnvVarMocker;
-  readonly endpointManager: InternetEndpointManager;
+  readonly endpoint: InternetEndpoint;
 }
 
 export function makeTestServer(
@@ -24,7 +24,7 @@ export function makeTestServer(
   const envVarMocker = configureMockEnvVars(envVars);
   const mockLogging = makeMockLogging();
   const getConnection = setUpTestDbConnection();
-  const getEndpointManager = mockInternetEndpoint(getConnection);
+  const getEndpoint = mockInternetEndpoint(getConnection);
 
   let server: FastifyInstance;
   beforeEach(async () => {
@@ -40,6 +40,6 @@ export function makeTestServer(
     dbConnection: getConnection(),
     logs: mockLogging.logs,
     envVarMocker,
-    endpointManager: getEndpointManager(),
+    endpoint: getEndpoint(),
   });
 }
