+++ src/utilities/awala/InternetEndpoint.ts
@@ -1,22 +1,71 @@
 import {
+  derDeserializeRSAPublicKey,
   Endpoint,
+  getIdFromIdentityKey,
   InvalidMessageError,
   type KeyStoreSet,
+  MockCertificateStore,
+  MockPublicKeyStore,
   NodeConnectionParams,
   type Parcel,
   type SessionKey,
   SessionKeyPair,
 } from '@relaycorp/relaynet-core';
+import envVar from 'env-var';
+import type { Connection } from 'mongoose';
+import { initPrivateKeystoreFromEnv } from '@relaycorp/awala-keystore-cloud';
 
-import { type Config, ConfigKey } from '../config.js';
+import { Config, ConfigKey } from '../config.js';
+import { Kms } from '../kms/Kms.js';
+
+import { InternetPrivateEndpointChannel } from './InternetPrivateEndpointChannel.js';
+
+function initKeyStoreSet(dbConnection: Connection): KeyStoreSet {
+  const privateKeyStoreAdapter = envVar.get('PRIVATE_KEY_STORE_ADAPTER').required().asString();
+  const privateKeyStore = initPrivateKeystoreFromEnv(privateKeyStoreAdapter, dbConnection);
+  return {
+    certificateStore: new MockCertificateStore(),
+    publicKeyStore: new MockPublicKeyStore(),
+    privateKeyStore,
+  };
+}
+
+async function retrieveIdentityPrivateKeyRef(): Promise<CryptoKey> {
+  const activeIdKeyRef = envVar.get('ACTIVE_ID_KEY_REF').required().asString();
+  const kms = await Kms.init();
+  return kms.retrievePrivateKeyByRef(Buffer.from(activeIdKeyRef));
+}
+
+async function getIdentityPublicKey() {
+  const activeIdPublicKeyBase64 = envVar.get('ACTIVE_ID_PUBLIC_KEY').required().asString();
+  return derDeserializeRSAPublicKey(Buffer.from(activeIdPublicKeyBase64, 'base64'));
+}
 
 export class InternetEndpoint extends Endpoint {
+  public static async getActive(dbConnection: Connection): Promise<InternetEndpoint> {
+    const privateKey = await retrieveIdentityPrivateKeyRef();
+    const publicKey = await getIdentityPublicKey();
+    const endpointId = await getIdFromIdentityKey(publicKey);
+    const internetAddress = envVar.get('INTERNET_ADDRESS').required().asString();
+    const keyStoreSet = initKeyStoreSet(dbConnection);
+    const config = new Config(dbConnection);
+    return new InternetEndpoint(
+      endpointId,
+      internetAddress,
+      { privateKey, publicKey },
+      keyStoreSet,
+      config,
+    );
+  }
+
+  protected readonly channelConstructor = InternetPrivateEndpointChannel;
+
   public constructor(
     id: string,
     public readonly internetAddress: string,
     identityKeyPair: CryptoKeyPair,
     keyStores: KeyStoreSet,
-    public readonly config: Config,
+    protected readonly config: Config,
   ) {
     super(id, identityKeyPair, keyStores, {});
   }
