+++ src/utilities/awala/InternetPrivateEndpointChannel.ts
@@ -1,0 +1,3 @@
+import { Channel, type ServiceMessage } from '@relaycorp/relaynet-core';
+
+export class InternetPrivateEndpointChannel extends Channel<ServiceMessage, string> {}
