+++ src/utilities/fastify/plugins/fastifyActiveEndpoint.spec.ts
@@ -5,8 +5,8 @@ import fastify from 'fastify';
 import { configureMockEnvVars } from '../../../testUtils/envVars.js';
 import { mockInternetEndpoint } from '../../../testUtils/awala/mockInternetEndpoint.js';
 import { MONGODB_URI, setUpTestDbConnection } from '../../../testUtils/db.js';
-import type { InternetEndpointManager } from '../../awala/InternetEndpointManager.js';
 import { getPromiseRejection } from '../../../testUtils/jest.js';
+import type { InternetEndpoint } from '../../awala/InternetEndpoint.js';
 
 import fastifyActiveEndpoint from './fastifyActiveEndpoint.js';
 import fastifyMongoose from './fastifyMongoose.js';
@@ -15,15 +15,14 @@ describe('fastifyActiveEndpoint', () => {
   configureMockEnvVars({ MONGODB_URI });
 
   const getConnection = setUpTestDbConnection();
-  const getInternetEndpointManager = mockInternetEndpoint(getConnection);
-  let internetEndpointManager: InternetEndpointManager;
 
+  const getInternetEndpoint = mockInternetEndpoint(getConnection);
+  let internetEndpoint: InternetEndpoint;
   beforeEach(() => {
-    internetEndpointManager = getInternetEndpointManager();
+    internetEndpoint = getInternetEndpoint();
   });
 
   test('Fastify should be decorated with get active endpoint function', async () => {
-    const internetEndpoint = await internetEndpointManager.getActiveEndpoint();
     const app = fastify();
     await app.register(fastifyMongoose);
     await app.register(fastifyActiveEndpoint);
